--- main.py
@@ -1,12 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-
-@app.route('/')
-def index():
-    return '<h1>Hello Chrome!</h1>'
-
-
-if __name__ == '__main__':
-    app.run(port=8080)
