+++ static/css/index.css
@@ -1,0 +1,24 @@
+.header {
+    width: 100%;
+    height: 60px;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, .1);
+    background-color: floralwhite;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.mnu_top {
+    list-style-type: none;
+    padding-left: 0;
+    margin-bottom: 0;
+    
+}
+
+.mnu_top li {
+    display: inline-block;
+}
+
+.logo {
+    margin-left: 50px;
+}
