+++ templates/base.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>{{title}}</title>
+    <!--Linked css-->
+    <!-- <link href={{css|safe}} rel="stylesheet"> -->
+    <link href="/static/css/index.css" rel="stylesheet">
+    <!--Boostrap css-->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous">
+</head>
+<body>
+    <header class="header">
+        <a href="/" class="logo"><img src="" alt="">Boosting</a>
+        <ul class="mnu_top">
+            <li><a href="/">Игры</a></li>
+            <li><a href="/">Заказы</a></li>
+            <li><a href="/">Бустеры</a></li>
+        </ul>
+        <div class="login_top">
+            <button class="btn btn-primary">Личный кабинет</button>
+        </div>
+    </header>
+</body>
+</html>
