+++ templates/index.html
@@ -1,0 +1,4 @@
+{% extends "base.html"%}
+{start block}
+
+{endblock}
